frontend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
jinja2==3.1.2
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
httpx==0.24.0
pydantic==1.10.7
python-dotenv==1.0.0
requests==2.28.2
aiofiles==23.1.0
plotly==5.14.1
pandas==2.0.1
pymongo==4.3.3
